done with the add item button, need to check if cache clears
