+++ components/TraySizeSelector/TraySizeSelector.behavior.comp.cy.js
@@ -1,4 +1,5 @@
 import TraySizeSelector from '@comps/TraySizeSelector/TraySizeSelector.vue';
+import * as farmosUtil from '@libs/farmosUtil/farmosUtil.js';
 
 describe('Test the TraySizeSelector behaviors', () => {
   beforeEach(() => {
@@ -11,7 +12,7 @@ describe('Test the TraySizeSelector behaviors', () => {
     cy.saveSessionStorage();
   });
 
-  it('Clicking add tray size button goes to the add tray size form', () => {
+  it('Clicking add tray size button goes to the add tray size form and clears the tray_sizes cache', () => {
     const readySpy = cy.spy().as('readySpy');
 
     cy.intercept('GET', '**/taxonomy/manage/tray_size/add', {
@@ -27,9 +28,17 @@ describe('Test the TraySizeSelector behaviors', () => {
       cy.get('@readySpy')
         .should('have.been.calledOnce')
         .then(() => {
+          expect(farmosUtil.getFromGlobalVariableCache('tray_sizes')).to.not.be
+            .null;
           cy.get('[data-cy="selector-add-button"]').should('exist');
           cy.get('[data-cy="selector-add-button"]').click();
-          cy.wait('@urlIntercept').its('response.statusCode').should('eq', 200);
+          cy.wait('@urlIntercept')
+            .its('response.statusCode')
+            .should('eq', 200)
+            .then(() => {
+              expect(farmosUtil.getFromGlobalVariableCache('tray_sizes')).to.be
+                .null;
+            });
         });
     });
   });
